Coinbase pagination  (#149)
* Adding IasyncEnumerable to retrieve transactions

* CoinbaseClient Modification

* Update src/Trakx.IndiceManager.Server/Data/CoinbaseTransactionListener.cs

* suggestions

* implementing IAsyncEnumerable in coinbaseClient

* Apply Changes after Matthieu's review and new tests

* quick suggestion for simplification 🤔

* Apply Matthieu's suggestion

* Change visibility of FromPagedResponse

* Modifying tests of CoinbaseTransactionListener

* Quick change in paginationOptions

* Codacy review

* renaming and testing around Coinbase Transactions

* remove LookingForCompletePage

* better endpoint method signature

* cleanup, big refactor to make CoinbaseClient testable

* codacy fixes

+++ src/Trakx.Coinbase.Custody.Client/Models/Currency.cs
@@ -4,14 +4,16 @@ namespace Trakx.Coinbase.Custody.Client.Models
 {
     public class Currency
     {
+        #nullable disable
         [JsonPropertyName("symbol")]
         public string Symbol { get; set; }
 
         [JsonPropertyName("name")]
         public string Name { get; set; }
+        #nullable restore
 
         [JsonPropertyName("decimals")]
-        public long Decimals { get; set; }
+        public ushort Decimals { get; set; }
     }
 }
